feat(channels): shared ChannelHeader — clickable name + tabs on every view
Extract a ChannelHeader (clickable "# channel" → channel home, plus the tab
strip) and use it across Home / Recents / Artifacts / CONTEXT.md, so the nav
tabs stay in view on the CONTEXT.md page too and the channel name navigates
home like the sidebar row.

Generated-By: PostHog Code
Task-Id: 50785ddb-a7ee-4710-bf17-1dd6205b6121
